remove circular dep in openai llm + agent
