+++ llama-index-integrations/agent/llama-index-agent-openai/pyproject.toml
@@ -28,13 +28,14 @@ exclude = ["**/BUILD"]
 license = "MIT"
 name = "llama-index-agent-openai"
 readme = "README.md"
-version = "0.3.1"
+version = "0.3.2"
 
 [tool.poetry.dependencies]
 python = ">=3.8.1,<4.0"
-llama-index-llms-openai = "^0.2.0"
 openai = ">=1.14.0"
+llama-index-llms-openai = "^0.2.0"
 llama-index-core = "^0.11.0"
+llama-index-utils-openai = "^0.1.0"
 
 [tool.poetry.group.dev.dependencies]
 ipython = "8.10.0"
