+++ llama-index-utils/llama-index-utils-openai/BUILD
@@ -1,0 +1,7 @@
+poetry_requirements(
+    name="poetry",
+)
+
+python_requirements(
+    name="reqs",
+)
