+++ llama-index-utils/llama-index-utils-openai/README.md
@@ -1,0 +1,3 @@
+# LlamaIndex OpenAI Utils
+
+The `llama-index-utils-openai` library contains common OpenAI-related utilities used across various packages.
