+++ llama-index-utils/llama-index-utils-openai/llama_index/utils/openai/__init__.py
@@ -1,0 +1,5 @@
+from llama_index.utils.openai.tools import (
+    resolve_tool_choice,
+)
+
+__all__ = ["resolve_tool_choice"]
