+++ llama-index-utils/llama-index-utils-openai/llama_index/utils/openai/tools.py
@@ -1,0 +1,14 @@
+"""Utils for OpenAI agent."""
+
+from typing import Union
+
+
+def resolve_tool_choice(tool_choice: Union[str, dict] = "auto") -> Union[str, dict]:
+    """Resolve tool choice.
+
+    If tool_choice is a function name string, return the appropriate dict.
+    """
+    if isinstance(tool_choice, str) and tool_choice not in ["none", "auto", "required"]:
+        return {"type": "function", "function": {"name": tool_choice}}
+
+    return tool_choice
